refactor to allow multiple crashsite presets
