+++ map_gen/maps/crash_site_desert.lua
@@ -1,0 +1,1 @@
+return require 'map_gen.maps.crash_site.presets.desert'
